Backend baselayout (#4)
### What?
Adds an initial code structure for backend to start building on.
Implements sections 15 and 44 as an example.

### Why?
Such that we don't end up in a huge chaos once everyone starts coding
their own sections and features.

### How?
Implements section as individual components/classes. Frontend receives a
generic visualization-template and all logic has to be implemented in
backend. This isn't too simple but very powerful.

### Testing?
Sure :) Btw it makes sense for everyone to use the `TestBase` class such
that the spring context is only initialized once.

### Anything Else?
- Adapted issue and PR template to be more expressive
- Added CODEOWNERS as discussed