Load section YAML files (#342)
### What?

This PR implements the loading of the YAML according to the example [in
the wiki](https://github.com/keykey7/fabbwled/wiki/Backend-conventions).

### Why?

This is the first step towards implementing the YAML sections fully. It
also makes it so that the sections are validated at startup, meaning
people can actually check whether they're doing it correctly.

### How?

We load all sections (the files are hardcoded) from the resources and
parse them as YAML. We then convert the "raw" YAML structures into
semantic classes for the actions and conditions, [validating as we
go](https://lexi-lambda.github.io/blog/2019/11/05/parse-don-t-validate/).

### Testing?

Start the backend and see what happens. I don't have tests yet, I will
add them to this PR later.

### Anything Else?

[//]: # (for a sample see:
https://www.pullrequest.com/blog/writing-a-great-pull-request-description/)

---------

Co-authored-by: Lion-alt <81552607+Lion-alt@users.noreply.github.com>