Playable CLI Client (#485)
### What?
A playable CLI client.
Renders text with a minimum of console styling and presents pickable
options as a simple AWT dialog popup.

### Why?
To allow easy testing of sections without a FE.

### How?
Implemented as a junit test (disabled if headless).
The AWT dialog sucks, but stdin is intercepted by Gradle so this is an
ugly solution to allow any blocking user interaction during tests. A
restful client would have been a more complicated alternative, but
doesn't solve the issue of blocking user interaction.